requirements.txt
# Web Framework
Flask==3.0.3

# Database & ORM
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0

# Security & Authentication
bcrypt==4.1.2
Flask-Login==0.6.3

# Environment & Utilities
python-dotenv==1.0.1
email-validator==2.1.1